// src/app/model/user-model.model.ts
export class UserModel {
    id?: number = 0;
    fullname: string = '';
    email: string = '';
    password: string = '';
    agreeterms: boolean = true;    
    authtoken: string = '';
    ipaddress: string = '';
    activestatus: boolean = false;
    created_at: string = '';
    modified_at: string = '';    
}